+++ global.h
@@ -47,3 +47,9 @@ void mh_sockhexdump(int fd, void *p, size_t len);
 int or20client(int fd);
 
 int del_client(int fd);
+
+int mochad_diag_hello(int fd);
+int mochad_diag_capabilities(int fd);
+int mochad_diag_health(int fd);
+int mochad_diag_clients(int fd);
+int mochad_diag_version(int fd);